# Conflicts:
#	smartbutlerquery.sql
#	src/main/java/com/fms/smartbutler/config/SecurityConfig.java
#	src/main/java/com/fms/smartbutler/repository/ClaimRepository.java
#	src/main/java/com/fms/smartbutler/repository/EstimateRepository.java
#	src/main/java/com/fms/smartbutler/service/ClaimService.java
#	src/main/java/com/fms/smartbutler/service/ResidentService.java

+++ src/main/java/com/fms/smartbutler/dto/ResidentDTO.java
@@ -1,5 +1,10 @@
 package com.fms.smartbutler.dto;
 
+/**
+* @author 엄다빈
+* @editDate 2024-02-01 ~ 2024-02-02
+*/
+
 import com.fms.smartbutler.domain.Build;
 import com.fms.smartbutler.domain.Users;
 
@@ -17,7 +22,7 @@ public class ResidentDTO {
 	private String buildName;
 	private Long userId;
 	private String userName;
-	private Long residentNum;
+	private int residentNum;
 	private String phone;
 	private int entered;
 	private String fromDate;
